requirements.txt
anthropic~=0.52.0
beautifulsoup4~=4.13.3
chardet~=5.2.0
matplotlib~=3.9.4
nltk~=3.9.1
numpy~=2.2.3
ollama~=0.4.7
openai~=1.63.2
pandas~=2.2.3
py3langid~=0.3.0
pydantic~=2.10.6
python-dotenv~=1.0.1
readability~=0.3.2
scikit-learn~=1.6.1
scipy~=1.15.2
selenium~=4.28.1
sentence-transformers~=3.4.1
tiktoken~=0.9.0
torch~=2.7.0
tqdm~=4.67.1
umap~=0.1.1
websockets~=15.0.1
wsproto~=1.2.0